+++ lib/src/components/JournalOverviewBadge.vue
